@@ -20,19 +20,19 @@
     "@apollo/client": "^3.5.10",
     "@shopify/app-bridge": "^2.0.22",
     "@shopify/app-bridge-react": "^2.0.24",
-    "@shopify/app-bridge-utils": "^2.0.23",
-    "@shopify/polaris": "^9.2.2",
+    "@shopify/app-bridge-utils": "^2.0.24",
+    "@shopify/polaris": "^9.4.0",
     "@shopify/shopify-api": "^3.0.0",
-    "@vitejs/plugin-react": "1.2.0",
+    "@vitejs/plugin-react": "1.3.0",
     "compression": "^1.7.4",
     "cookie-parser": "^1.4.6",
     "cross-env": "^7.0.3",
     "dotenv": "^16.0.0",
     "express": "^4.17.3",
     "graphql": "^16.3.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "serve-static": "^1.14.1",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "serve-static": "^1.15.0",
     "vite": "^2.9.1"
   },
   "devDependencies": {
@@ -41,6 +41,6 @@
     "prettier": "^2.6.2",
     "pretty-quick": "^3.1.3",
     "supertest": "^6.2.2",
-    "vitest": "^0.9.2"
+    "vitest": "^0.9.3"
   }
 }
